+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/DataFactory.java
@@ -59,15 +59,10 @@ import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.TextFormatter;
 import com.liferay.portal.model.impl.ResourcePermissionModelImpl;
-import com.liferay.wiki.constants.WikiPageConstants;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiNodeModel;
 import com.liferay.wiki.model.WikiPage;
 import com.liferay.wiki.model.WikiPageModel;
-import com.liferay.wiki.model.WikiPageResourceModel;
-import com.liferay.wiki.model.impl.WikiNodeModelImpl;
-import com.liferay.wiki.model.impl.WikiPageModelImpl;
-import com.liferay.wiki.model.impl.WikiPageResourceModelImpl;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
@@ -139,12 +134,11 @@ public class DataFactory extends BaseDDMDataFactory {
 		else if (name.equals("userDataFactory")) {
 			return UserDataFactory.getInstance();
 		}
+		else if (name.equals("wikiDataFactory")) {
+			return WikiDataFactory.getInstance();
+		}
 
 		return ClassNameDataFactory.getInstance();
-	}
-
-	public int getMaxWikiPageCommentCount() {
-		return BenchmarksPropsValues.MAX_WIKI_PAGE_COMMENT_COUNT;
 	}
 
 	public ResourcePermissionModel newCommerceCatalogResourcePermissionModel(
@@ -358,51 +352,6 @@ public class DataFactory extends BaseDDMDataFactory {
 			String.valueOf(wikiPageModel.getResourcePrimKey()), SAMPLE_USER_ID);
 	}
 
-	public List<WikiNodeModel> newWikiNodeModels(long groupId) {
-		List<WikiNodeModel> wikiNodeModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_WIKI_NODE_COUNT);
-
-		for (int i = 1; i <= BenchmarksPropsValues.MAX_WIKI_NODE_COUNT; i++) {
-			wikiNodeModels.add(newWikiNodeModel(groupId, i));
-		}
-
-		return wikiNodeModels;
-	}
-
-	public List<WikiPageModel> newWikiPageModels(WikiNodeModel wikiNodeModel) {
-		List<WikiPageModel> wikiPageModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_WIKI_PAGE_COUNT);
-
-		for (int i = 1; i <= BenchmarksPropsValues.MAX_WIKI_PAGE_COUNT; i++) {
-			wikiPageModels.add(newWikiPageModel(wikiNodeModel, i));
-		}
-
-		return wikiPageModels;
-	}
-
-	public WikiPageResourceModel newWikiPageResourceModel(
-		WikiPageModel wikiPageModel) {
-
-		WikiPageResourceModel wikiPageResourceModel =
-			new WikiPageResourceModelImpl();
-
-		// UUID
-
-		wikiPageResourceModel.setUuid(SequentialUUID.generate());
-
-		// PK fields
-
-		wikiPageResourceModel.setResourcePrimKey(
-			wikiPageModel.getResourcePrimKey());
-
-		// Other fields
-
-		wikiPageResourceModel.setNodeId(wikiPageModel.getNodeId());
-		wikiPageResourceModel.setTitle(wikiPageModel.getTitle());
-
-		return wikiPageResourceModel;
-	}
-
 	public String toInsertSQL(BaseModel<?> baseModel) {
 		try {
 			StringBundler sb = new StringBundler();
@@ -536,84 +485,6 @@ public class DataFactory extends BaseDDMDataFactory {
 			newResourcePermissionModel(name, primKey, SITE_MEMBER_ROLE_ID, 0));
 
 		return resourcePermissionModels;
-	}
-
-	protected WikiNodeModel newWikiNodeModel(long groupId, int index) {
-		WikiNodeModel wikiNodeModel = new WikiNodeModelImpl();
-
-		// UUID
-
-		wikiNodeModel.setUuid(SequentialUUID.generate());
-
-		// PK fields
-
-		wikiNodeModel.setNodeId(counter.get());
-
-		// Group instance
-
-		wikiNodeModel.setGroupId(groupId);
-
-		// Audit fields
-
-		wikiNodeModel.setCompanyId(COMPANY_ID);
-		wikiNodeModel.setUserId(SAMPLE_USER_ID);
-		wikiNodeModel.setUserName(SAMPLE_USER_NAME);
-		wikiNodeModel.setCreateDate(new Date());
-		wikiNodeModel.setModifiedDate(new Date());
-
-		// Other fields
-
-		wikiNodeModel.setName("Test Node " + index);
-		wikiNodeModel.setLastPostDate(new Date());
-		wikiNodeModel.setLastPublishDate(new Date());
-		wikiNodeModel.setStatusDate(new Date());
-
-		return wikiNodeModel;
-	}
-
-	protected WikiPageModel newWikiPageModel(
-		WikiNodeModel wikiNodeModel, int index) {
-
-		WikiPageModel wikiPageModel = new WikiPageModelImpl();
-
-		// UUID
-
-		wikiPageModel.setUuid(SequentialUUID.generate());
-
-		// PK fields
-
-		wikiPageModel.setPageId(counter.get());
-
-		// Resource
-
-		wikiPageModel.setResourcePrimKey(counter.get());
-
-		// Group instance
-
-		wikiPageModel.setGroupId(wikiNodeModel.getGroupId());
-
-		// Audit fields
-
-		wikiPageModel.setCompanyId(COMPANY_ID);
-		wikiPageModel.setUserId(SAMPLE_USER_ID);
-		wikiPageModel.setUserName(SAMPLE_USER_NAME);
-		wikiPageModel.setCreateDate(new Date());
-		wikiPageModel.setModifiedDate(new Date());
-
-		// Other fields
-
-		wikiPageModel.setNodeId(wikiNodeModel.getNodeId());
-		wikiPageModel.setTitle("Test Page " + index);
-		wikiPageModel.setVersion(WikiPageConstants.VERSION_DEFAULT);
-		wikiPageModel.setContent(
-			StringBundler.concat(
-				"This is Test Page ", index, " of ", wikiNodeModel.getName(),
-				"."));
-		wikiPageModel.setFormat("creole");
-		wikiPageModel.setHead(true);
-		wikiPageModel.setLastPublishDate(new Date());
-
-		return wikiPageModel;
 	}
 
 	protected void toInsertSQL(StringBundler sb, BaseModel<?> baseModel) {
